fix: meltano nix build (#2373)
* fix: properly bypass fasteners test

* refactor: simplify fasteners fix

* fix: skip 'portalocker' tests

* refactor: move patch to meltano.nix

* revert: move patch to meltano.nix

This reverts commit f030ca38b3904a5022f365ad22badf5c6351357d. I could
not find any evidence that this change was affecting the derivation
ibeing built after trying to inspect the build, log traces, and observe
the artefacts produced.

* chore: remove portalocker override

This override still doesn't get to the derivation setting since it is
being re-enabled internally in a later phase.

* fix: switch from packages to interpreter for override

This is to fix the issue where portalocker derivation was not being
changed with the previous approach.

* fix: add debugpy to disableTests

* refactor: use 'genAttrs' to iterate of pkgs

* core: fix flake.nix format

* chore: re-add virtualenv in meltano.nix